+++ MSGraph/changelog.md
@@ -2,6 +2,7 @@
 # 1.2.8.3
 - Upd: command New-MgaMailMessage, Send-MgaMailMessage, Set-MgaMailMessage
     - Update parameter "ToRecipients" -> add parameter alias names "To" and "Reciepients"
+    - some internal code refactoring
 - Upd: command Send-MgaMailMessage
     - Add parameter set, to send new mail directly from command without need to use New-MgaMailMessage before Send-MgaMailMessage